Restructured file layout Moved around files to better represent route and api layout.
