+++ packages/frontend/src/components/groups/groupDetails.tsx
